refactor(feature): refactor badge-badge feature - badge-badge feature의 일부 파일 패키지 경로를 업데이트 합니다.
